same merge, but added files this time

+++ lib/python/Components/Lcd.py
@@ -103,10 +103,11 @@ class LCD:
 		return eDBoxLCD.getInstance().isOled()
 
 	def setMode(self, value):
-		print 'setLCDMode',value
-		f = open("/proc/stb/lcd/show_symbols", "w")
-		f.write(value)
-		f.close()
+		if fileExists("/proc/stb/lcd/show_symbols"):
+			print 'setLCDMode',value
+			f = open("/proc/stb/lcd/show_symbols", "w")
+			f.write(value)
+			f.close()
 		if config.lcd.mode.getValue() == "0":
 			if fileExists("/proc/stb/lcd/symbol_hdd"):
 				f = open("/proc/stb/lcd/symbol_hdd", "w")
@@ -194,6 +195,33 @@ def InitLcd():
 	SystemInfo["Display"] = detected
 	config.lcd = ConfigSubsection();
 	if detected:
+		if fileExists("/proc/stb/lcd/mode"):
+			f = open("/proc/stb/lcd/mode", "r")
+			can_lcdmodechecking = f.read().strip().split(" ")
+			f.close()
+		else:
+			can_lcdmodechecking = False
+
+		SystemInfo["LCDMiniTV"] = can_lcdmodechecking
+
+		if can_lcdmodechecking:
+			def setLCDModeMinitTV(configElement):
+				try:
+					f = open("/proc/stb/lcd/mode", "w")
+					f.write(configElement.value)
+					f.close()
+				except:
+					pass
+			config.lcd.modeminitv = ConfigSelection(choices={
+					"0": _("normal"),
+					"1": _("MiniTV"),
+					"2": _("OSD"),
+					"3": _("MiniTV with OSD")},
+					default = "0")
+			config.lcd.modeminitv.addNotifier(setLCDModeMinitTV)
+		else:
+			config.lcd.modeminitv = ConfigNothing()
+
 		def setLCDbright(configElement):
 			try:
 				ilcd.setBright(configElement.getValue());
@@ -283,7 +311,7 @@ def InitLcd():
 			config.lcd.contrast = ConfigNothing()
 			standby_default = 1
 
-		if getBoxType() == 'ebox5000':
+		if getBoxType() == 'ebox5000' or getBoxType() == 'ebox5100':
 			config.lcd.standby = ConfigSlider(default=standby_default, limits=(0, 4))
 			config.lcd.bright = ConfigSlider(default=4, limits=(0, 4))
 		else:
@@ -301,7 +329,7 @@ def InitLcd():
 		config.lcd.flip = ConfigYesNo(default=False)
 		config.lcd.flip.addNotifier(setLCDflipped);
 
-		if getBoxType() == 'ebox5000':
+		if getBoxType() == 'ebox5000' or getBoxType() == 'ebox5100':
 			config.lcd.scrollspeed = ConfigSlider(default = 150, increment = 10, limits = (0, 500))
 			config.lcd.scrollspeed.addNotifier(setLCDscrollspeed);
 			config.lcd.repeat = ConfigSelection([("0", _("None")), ("1", _("1X")), ("2", _("2X")), ("3", _("3X")), ("4", _("4X")), ("500", _("Continues"))], "3")
